refactor: [#7] rename folder
