+++ .github/actions/git-queue/__tests__/main.test.ts
@@ -2,7 +2,7 @@ import * as process from 'process';
 import * as cp from 'child_process';
 import * as path from 'path';
 import {expect, test} from '@jest/globals';
-import {createTmpDir, dummyPayload, gitLogForLatestCommit, newSimpleGit} from '../src/__helpers__/helpers';
+import {createTmpDir, dummyPayload, gitLogForLatestCommit, newSimpleGit} from '../src/__tests__/helpers';
 
 function executeAction(env) {
   const np = process.execPath;
