+++ .github/actions/git-queue/package.json
@@ -6,7 +6,7 @@
     "build": "tsc && ncc build",
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
-    "test": "yarn run build && jest --coverage",
+    "test": "jest --coverage",
     "pre-checkin": "yarn run format && yarn run build"
   },
   "repository": {
